refactor: created di-container for extracting initialization. Modified routes
