+++ src/api/routes.ts
@@ -1,31 +1,7 @@
 import express from 'express'
-import { PrismaClient } from '@prisma/client'
-import { UserPrismaRepository } from '../Contexts/user/infrastructure/repositories/user-prisma.repository'
-
-import { UsersController } from './controllers/users.controller'
-import {
-  GetAllUsersService,
-  GetUserByIdService,
-  CreateUserService,
-  UpdateUserService,
-  DeleteUserService,
-} from '../Contexts/user/application'
+import { usersController } from './config/di-container'
 
 export const userRouter = express.Router()
-const prismaClient = new PrismaClient()
-const userPrismaRepository = new UserPrismaRepository(prismaClient)
-const getAllUsersService = new GetAllUsersService(userPrismaRepository)
-const getUserByIdService = new GetUserByIdService(userPrismaRepository)
-const createUserService = new CreateUserService(userPrismaRepository)
-const updateUserService = new UpdateUserService(userPrismaRepository)
-const deleteUserService = new DeleteUserService(userPrismaRepository)
-const usersController = new UsersController(
-  getAllUsersService,
-  getUserByIdService,
-  createUserService,
-  updateUserService,
-  deleteUserService,
-)
 
 userRouter.get('/', usersController.getUsers)
